Move TestDummyValidator/Wrapper to toplevel classes for re-use in Receiver tests
